api/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
decorator==5.1.1
dnspython==2.7.0
email_validator==2.2.0
executing==2.1.0
fastapi==0.115.2
h11==0.14.0
idna==3.10
ipython==8.28.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
MarkupSafe==3.0.1
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
packaging==24.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
PyJWT==2.9.0
Pygments==2.18.0
pymongo==4.10.1
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.40.0
traitlets==5.14.3
typing_extensions==4.12.2
uvicorn==0.32.0
wcwidth==0.2.13
Werkzeug==3.0.4
